fix(tcp_sender): use correct data type

+++ libsponge/tcp_sender.cc
@@ -2,6 +2,7 @@
 
 #include "tcp_config.hh"
 
+#include <cstddef>
 #include <random>
 
 // Dummy implementation of a TCP sender
@@ -10,7 +11,7 @@
 // automated checks run by `make check_lab3`.
 
 template <typename... Targs>
-void DUMMY_CODE(Targs &&... /* unused */) {}
+void DUMMY_CODE(Targs &&.../* unused */) {}
 
 using namespace std;
 
@@ -22,7 +23,7 @@ TCPSender::TCPSender(const size_t capacity, const uint16_t retx_timeout, const s
     , _initial_retransmission_timeout{retx_timeout}
     , _stream(capacity) {}
 
-uint64_t TCPSender::bytes_in_flight() const { return {}; }
+size_t TCPSender::bytes_in_flight() const { return {}; }
 
 void TCPSender::fill_window() {}
